requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
mysqlclient==2.2.0
redis==5.0.1
influxdb-client==1.38.0

# 数据处理
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# HTTP客户端
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# 认证和安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.8.0

# 任务调度
apscheduler==3.10.4
celery==5.3.4

# 配置管理
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 日志
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 开发工具
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 金融数据
yfinance==0.2.28
ta-lib==0.4.28
quandl==3.7.0

# 机器学习
scikit-learn==1.3.2
lightgbm==4.1.0
xgboost==2.0.2

# 图表和可视化
plotly==5.17.0
matplotlib==3.8.2

# 工具库
click==8.1.7
tqdm==4.66.1
pytz==2023.3
python-dateutil==2.8.2

# 监控
prometheus-client==0.19.0

# 缓存
cachetools==5.3.2

# 异步支持
aiofiles==23.2.1
aioredis==2.0.1

# 数据验证
cerberus==1.3.5
marshmallow==3.20.1

# 网络
websockets==12.0

# 系统监控
psutil==5.9.6